myproject/requirements.txt
Django>=5.0
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.2.2
psycopg2-binary
gunicorn
django-cors-headers
